src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.6
openai==1.78.1
opencv-python==4.11.0.86
packaging==25.0
pdfminer.six==20250506
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
pytesseract==0.3.13
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wikipedia==1.4.0